fix: move the sidebar toggle into the sidebar footer next to the user menu
