+++ apps/ui/src/components/sidebar-header/index.tsx
@@ -4,15 +4,10 @@ import { menu, plus } from '@wordpress/icons';
 import { IconButton } from '@wordpress/ui';
 import { useConnector } from '@/data/core';
 import { useTrafficLightSpace } from '@/hooks/use-traffic-light-space';
-import { drawerIcon } from '@/lib/icons';
 import styles from './style.module.css';
 import type { MouseEvent } from 'react';
 
-type Props = {
-	onToggleSidebar: () => void;
-};
-
-export function SidebarHeader( { onToggleSidebar }: Props ) {
+export function SidebarHeader() {
 	const reserveTrafficLightSpace = useTrafficLightSpace();
 	const navigate = useNavigate();
 	const connector = useConnector();
@@ -41,14 +36,6 @@ export function SidebarHeader( { onToggleSidebar }: Props ) {
 					icon={ plus }
 					label={ __( 'Add site' ) }
 					onClick={ () => void navigate( { to: '/onboarding' } ) }
-				/>
-				<IconButton
-					variant="minimal"
-					tone="neutral"
-					size="small"
-					icon={ drawerIcon }
-					label={ __( 'Hide sidebar' ) }
-					onClick={ onToggleSidebar }
 				/>
 			</div>
 		</div>
