+++ apps/ui/src/components/sidebar-layout/index.test.tsx
@@ -5,9 +5,7 @@ import { SidebarLayout } from './index';
 import type { ReactNode } from 'react';
 
 vi.mock( '@/components/sidebar-header', () => ( {
-	SidebarHeader: ( { onToggleSidebar }: { onToggleSidebar: () => void } ) => (
-		<button onClick={ onToggleSidebar }>Hide sidebar</button>
-	),
+	SidebarHeader: () => null,
 } ) );
 
 vi.mock( '@/components/site-list', () => ( {
@@ -15,7 +13,9 @@ vi.mock( '@/components/site-list', () => ( {
 } ) );
 
 vi.mock( '@/components/user-menu', () => ( {
-	UserMenu: () => null,
+	UserMenu: ( { onToggleSidebar }: { onToggleSidebar: () => void } ) => (
+		<button onClick={ onToggleSidebar }>Hide sidebar</button>
+	),
 } ) );
 
 vi.mock( '@/data/core', () => ( {
