+++ apps/ui/src/components/user-menu/style.module.css
@@ -22,19 +22,22 @@
 
 .userName,
 .themeToggle,
+.sidebarToggle,
 .settingsButton {
 	color: var(--wpds-color-fg-content-neutral-weak);
 }
 
-.themeToggle {
+.sidebarToggle {
 	margin-inline-end: calc(-1 * var(--wpds-dimension-padding-sm));
 }
 
 .row:hover .userName,
 .row:hover .themeToggle,
+.row:hover .sidebarToggle,
 .row:hover .settingsButton,
 .row:focus-within .userName,
 .row:focus-within .themeToggle,
+.row:focus-within .sidebarToggle,
 .row:focus-within .settingsButton {
 	color: var(--wpds-color-fg-content-neutral);
 }
